Better timeout management
Start with a larger initial value, and use a larger increment.
Also rely more on the server responding with a 504, using the ajax
timeout as a last resort.
This avoids the UI Python library dealing with multiple concurrent
requests that end up being sent to the piju server.
Overall, this avoids lots of wasted network traffic, and a quicker
response to the browser.